requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
